+++ Hardware/Firmwares/RotatorEth-esp32/RemoteShutterClass.h
@@ -17,15 +17,15 @@ public:
 
 	// These aren't used by Rotator so why bother converting them to numeric values?
 	int		acceleration = 0;
-	double	elevation = 0;
+	float	elevation = 0.0f;
 	String	OpenError = "";
 	int		speed = 0;
 	bool	reversed = "";
 	int		stepsPerStroke = 0;
 	// ASCOM checks version and if it's blank then shutter doesn't exist
 	String	version = "";
-	double	volts = 0.0;
-	double	voltsCutOff = 0.0;
+	float	volts = 0.0f;
+	float	voltsCutOff = 0.0f;
 	unsigned int	watchdogInterval = 90; // set proper default.. just in case.
 	String  ssid = "RTIShutter";
 	String  lowVoltStateOrBadConditions = "";
